week8/day5/requirement.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.4
django-phonenumber-field==7.0.2
filelock==3.8.2
phonenumbers==8.13.4
phonenumberslite==8.13.4
platformdirs==2.6.0
psycopg2==2.9.5
PyMySQL==1.0.2
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.17.1
